Add support for { and } commands
This commit adds triangle compiler support for open curly brackets
{ and closed curly brackets } to be used as wrappers which allow
multiple commands to appear in one place inside triangle programs.
These operate the same way as BEGIN and END statements in triangle
programs.

This commit also adds a test in the TestScanner class of the
triangle compiler which tests that aforementioned new commands
work as expected.