Add loop...while contextual analysis & code generation
This commit adds the ability for the triangle compiler to perform contextual
analysis and code generation on triangle programs that use the
loop...while...do... command.

This commit also adds the ability for the triangle compiler to draw an
abstract syntax tree for triangle programs that use the
loop...while...do... command.