Add program statistics option to compiler
This commit adds the ProgramStatistics class which counts how many BinaryExpressions,
how many IfCommands, and how many WhileCommands there are in a given triangle program.

This commit also adds the stats argument to the Compiler class which is loaded using
the cli-parser library. When provided the comopiler generates statistics for the given
triangle program using the ProgramStatistics class.